chess_QA/synth-chess-dataset/requirements.txt
pytest>=8.0.0
datasets>=2.19.0
huggingface_hub>=0.24.0
Pillow>=10.0
